Feat: Migrate saveTokenAndSendEmail() from UserService to AuthFacade
