+++ src/main/java/com/koliving/api/user/IUserService.java
@@ -3,8 +3,6 @@ package com.koliving.api.user;
 
 public interface IUserService {
 
-    void saveTokenAndSendEmail(String mail);
-
     User signUp(String email);
 
     void setPassword(User user, String password);
