+++ src/main/java/com/koliving/api/user/UserService.java
@@ -1,10 +1,6 @@
 package com.koliving.api.user;
 
-import com.koliving.api.event.ConfirmationTokenCreatedEvent;
-import com.koliving.api.token.confirmation.ConfirmationToken;
-import com.koliving.api.token.confirmation.IConfirmationTokenService;
 import lombok.AllArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -18,8 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService implements IUserService, UserDetailsService {
 
     private final UserRepository userRepository;
-    private final IConfirmationTokenService confirmationTokenService;
-    private final ApplicationEventPublisher eventPublisher;
     private final PasswordEncoder passwordEncoder;
 
     @Override
@@ -27,13 +21,6 @@ public class UserService implements IUserService, UserDetailsService {
     public UserDetails loadUserByUsername(String email) {
         return userRepository.findByEmail(email).orElseThrow(() ->
                 new UsernameNotFoundException(String.format("User with email %s not found", email)));
-    }
-
-    @Override
-    public void saveTokenAndSendEmail(String email) {
-        ConfirmationToken newToken = confirmationTokenService.createToken(email);
-        ConfirmationToken savedToken = confirmationTokenService.saveToken(newToken);
-        eventPublisher.publishEvent(new ConfirmationTokenCreatedEvent(savedToken));
     }
 
     @Override
